improve heapsnapshot converter code style and readability
